requirements.txt
azure-iot-device==2.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
deprecation==2.1.0
idna==3.4
janus==1.0.0
packaging==23.0
paho-mqtt==1.6.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
requests-unixsocket==0.3.0
typing_extensions==4.5.0
urllib3==1.26.15
